feat: modify build config
