+++ CONTRIBUTING.md
@@ -4,7 +4,7 @@
 
 ```shell
 $ git clone https://github.com/MrSeaWave/chinese-holidays.git
-$ yarn
+$ pnpm i
 ```
 
 ### 调试
@@ -15,19 +15,19 @@ $ yarn
 
 ```shell
 
-$ npm run build
+$ pnpm run build
 
 ```
 
 #### Test
 
 ```shell
-$ npm run test
+$ pnpm run test
 ```
 
 ## Pull Request Etiquette
 
 - 必须严格按照 [commitlint](https://github.com/conventional-changelog/commitlint#what-is-commitlint) 格式提交代码
-- 请尽可能提供 [test](./__tests__) 文件
+- 请尽可能提供 test 文件
 - 请可能补充文档，[如何去使用](./README.md)
-- 请修改 CHANGELOG 文件，[如何去使用](./CHANGELOG.md)
+- 请修改 [CHANGELOG 文件](./CHANGELOG.md)
