+++ src/core/config.ts
@@ -1,5 +1,8 @@
 // 日期网址链接
 // export const BASE_URL = 'https://cdn.jsdelivr.net/gh/NateScarlet/holiday-cn@master';
+/**
+ * 中国地区 cdn 地址受到污染，因此使用 fastly.jsdelivr.net
+ */
 export const BASE_URL = 'https://fastly.jsdelivr.net/gh/NateScarlet/holiday-cn@master';
 
 export const WEEKS = ['星期日', '星期一', '星期二', '星期三', '星期四', '星期五', '星期六'];
